+++ operator/redisfailover/ensurer_test.go
@@ -50,13 +50,12 @@ func generateRF(enableExporter bool, bootstrapping bool) *redisfailoverv1.RedisF
 }
 
 func generateRFBootstrappingNode(bootstrapping bool) *redisfailoverv1.BootstrapSettings {
-	if bootstrapping {
-		return &redisfailoverv1.BootstrapSettings{
-			Host: "127.0.0.1",
-			Port: "6379",
-		}
+
+	return &redisfailoverv1.BootstrapSettings{
+		Host:    "127.0.0.1",
+		Port:    "6379",
+		Enabled: bootstrapping,
 	}
-	return nil
 }
 
 func TestEnsure(t *testing.T) {