+++ api/redisfailover/v1/types.go
@@ -175,7 +175,7 @@ type BootstrapSettings struct {
 	Host           string `json:"host,omitempty"`
 	Port           string `json:"port,omitempty"`
 	AllowSentinels bool   `json:"allowSentinels,omitempty"`
-        // +kubebuilder:default:true
+	// +kubebuilder:default=true
 	Enabled        bool   `json:"enabled,omitempty"`
 }