+++ operator/redisfailover/ensurer.go
@@ -46,16 +46,6 @@ func (w *RedisFailoverHandler) Ensure(rf *redisfailoverv1.RedisFailover, labels 
 		}
 	}
 
-	sentinelsAllowed := rf.SentinelsAllowed()
-	if sentinelsAllowed {
-		if err := w.rfService.EnsureSentinelService(rf, labels, or); err != nil {
-			return err
-		}
-		if err := w.rfService.EnsureSentinelConfigMap(rf, labels, or); err != nil {
-			return err
-		}
-	}
-
 	if err := w.rfService.EnsureRedisMasterService(rf, labels, or); err != nil {
 		return err
 	}
@@ -77,8 +67,21 @@ func (w *RedisFailoverHandler) Ensure(rf *redisfailoverv1.RedisFailover, labels 
 		return err
 	}
 
+	sentinelsAllowed := rf.SentinelsAllowed()
 	if sentinelsAllowed {
+
+		if err := w.rfService.EnsureSentinelService(rf, labels, or); err != nil {
+			return err
+		}
+		if err := w.rfService.EnsureSentinelConfigMap(rf, labels, or); err != nil {
+			return err
+		}
+
 		if err := w.rfService.EnsureSentinelDeployment(rf, labels, or); err != nil {
+			return err
+		}
+	} else {
+		if err := w.rfService.DestroySentinelResources(rf); err != nil {
 			return err
 		}
 	}