+++ operator/redisfailover/service/generator.go
@@ -169,7 +169,7 @@ func generateHAProxyConfigmap(rf *redisfailoverv1.RedisFailover, labels map[stri
 	port := rf.Spec.Redis.Port
 	haproxyCfg := fmt.Sprintf(`global
 	daemon
-	maxconn 256
+	maxconn 5000
 
 	defaults
 	mode tcp